Adjust global variables, reformat calcDisplay(), and add two new functions
The first change was to make 'counter' a global variable in order to reset it in clearDisplay() as the variable was enclosed in the function calcDisplay(). The second change was removing the decimal eventListener and giving it its own function as there was an issue with the '{once: true}' function and the decimal not being able to be used after the display and displayValue were cleared. Finally the clearDisplay() function was added to clear the display and the displayValue variable giving the user a fresh start.