Add function that executes the mathematical functions using operator buttons
The function doMath() adds eventListeners to the operator buttons which keep track of the ammount of times the buttons are pressed but only under the condition that displayValue.length is greater than 0. The event also stores the users first number in variable x, stores the selected operator in variable z, and stores the users second number in variable y. The event accounts for multiple operator button selections if the user clicks an operator they do not intend to use and selects their desired one while also accounting for the user doing multiple calculations before pressing equals.

There is also an eventListener for the equals button that stores the users second number and calculates the answer using the operate() function.