Add grid display to #keys, style #container and #display
Box-sizing was added to all elements. Container and display were re-sized and their boders were made slightly thicker along with similar border radius's added to both. #keys was turned into a grid container and it's height and width were adjusted to fit the new size of the container. All the button elements were assigned a gird area to fit the design in mind.